backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
scikit-learn==1.5.0
joblib==1.3.2
gunicorn==21.2.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
dnspython==2.4.2
pandas==2.1.0
opencv-python-headless
Pillow
numpy<2.0.0
